feat: eliminar archivos innecesarios y optimizar la configuración de la aplicación para el despliegue en Vercel
