+++ render.yaml
@@ -1,0 +1,18 @@
+services:
+  - type: web
+    name: gymai
+    env: python
+    plan: free
+    buildCommand: pip install -r requirements.txt && pip install gunicorn
+    startCommand: gunicorn -k uvicorn.workers.UvicornWorker -b 0.0.0.0:$PORT app.main:app
+    envVars:
+      - key: PYTHON_VERSION
+        value: 3.9.0
+      - key: GEMINI_API_KEY
+        sync: false
+      - key: DATABASE_URL
+        value: sqlite:///gym_ai.db
+      - key: SECRET_KEY
+        sync: false
+    healthCheckPath: /health
+    autoDeploy: true
