+++ requirements.txt
@@ -1,6 +1,7 @@
 # Dependencias básicas del framework
 fastapi==0.110.0
 uvicorn==0.27.1  # Actualizado por seguridad
+gunicorn==21.2.0  # Necesario para Render
 
 # Plantillas y procesamiento de formularios
 jinja2==3.1.3  # ADVERTENCIA: Actualizar cuando se publique la versión que solucione CVE-2024-22195/22196/22197
