--- vercel.json
@@ -1,75 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "version": 2,
-  "builds": [
-    {
-      "src": "app/vercel_app.py",
-      "use": "@vercel/python",
-      "config": {
-        "maxLambdaSize": "50mb",
-        "runtime": "python3.9"
-      }
-    },
-    {
-      "src": "build_vercel.sh",
-      "use": "@vercel/static-build",
-      "config": {
-        "distDir": "staticfiles"
-      }
-    }
-  ],
-  "routes": [
-    {
-      "src": "/ws/chat/[0-9]+",
-      "dest": "app/vercel_app.py",
-      "headers": {
-        "Connection": "upgrade",
-        "Upgrade": "websocket"
-      },
-      "continue": true
-    },
-    {
-      "src": "/api/routine/modify/[0-9]+",
-      "methods": ["POST"],
-      "dest": "app/vercel_app.py"
-    },
-    {
-      "src": "/api/modify_routine/[0-9]+",
-      "methods": ["POST"],
-      "dest": "app/vercel_app.py"
-    },
-    {
-      "src": "/static/(.*)",
-      "dest": "/static/$1",
-      "headers": {
-        "cache-control": "public, max-age=31536000, immutable"
-      }
-    },
-    {
-      "src": "/(styles\\.css)",
-      "dest": "/static/$1"
-    },
-    {
-      "src": "/css/(.*\\.css)",
-      "dest": "/static/css/$1"
-    },
-    {
-      "src": "/favicon.ico",
-      "dest": "/static/favicon.ico"
-    },
-    {
-      "src": "/api/(.*)",
-      "dest": "app/vercel_app.py"
-    },
-    {
-      "src": "/(.*)",
-      "dest": "app/vercel_app.py"
-    }
-  ],
-  "env": {
-    "PYTHONPATH": ".",
-    "DISABLE_COLLECTSTATIC": "1",
-    "SIMPLE_MODE": "true",
-    "DEBUG": "true"
-  }
-}
